# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "My API",
        "description": "All endpoints of backend requests are listed here\n",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://nodejs-homework-rest-api-7or0.onrender.com"
        }
    ],
...
    "paths": {
...
        "/courses": {
            "get": {
                "tags": ["Courses"],
                "summary": "Get all courses",
                "operationId": "getAllCourses",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/courses/:alias": {
            "get": {
                "tags": ["Courses"],
                "summary": "Get course by alias",
                "operationId": "getCourseByAlias",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/courses/products": {
            "post": {
                "tags": ["Courses"],
                "summary": "Get the course product",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Categories"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Course": {
                "type": "object",
                "properties": {
                    "_id": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "tags": {
                        "$ref": "#/components/schemas/properties/properties/Array"
                    },
                    "alias": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "title": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "category": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "tagsTitle": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "metaTitle": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "metaDescription": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "advantages": {
                        "$ref": "#/components/schemas/properties/properties/Advantages"
                    },
                    "seoText": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "faq": {
                        "$ref": "#/components/schemas/properties/properties/Advantages"
                    },
                    "salary": {
                        "$ref": "#/components/schemas/properties/properties/Salary"
                    }
                }
            },
            "Product": {
                "type": "object",
                "properties": {
                    "_id": {
                        "$ref": "#/components/schemas/properties/properties/Id"
                    },
                    "categories": {
                        "$ref": "#/components/schemas/properties/properties/Array"
                    },
                    "tags": {
                        "$ref": "#/components/schemas/properties/properties/Array"
                    },
                    "title": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "link": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "image": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "initialRating": {
                        "$ref": "#/components/schemas/properties/properties/Number"
                    },
                    "characteristics": {
                        "$ref": "#/components/schemas/properties/properties/Characteristics"
                    },
                    "description": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "price": {
                        "$ref": "#/components/schemas/properties/properties/Number"
                    },
                    "oldPrice": {
                        "$ref": "#/components/schemas/properties/properties/Number"
                    },
                    "credit": {
                        "$ref": "#/components/schemas/properties/properties/Number"
                    },
                    "advantages": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "createdAt": {
                        "$ref": "#/components/schemas/properties/properties/Date"
                    },
                    "updatedAt": {
                        "$ref": "#/components/schemas/properties/properties/Date"
                    },
                    "html": {
                        "$ref": "#/components/schemas/properties/properties/String"
                    },
                    "clicks": {
                        "$ref": "#/components/schemas/properties/properties/Number"
                    },
                    "reviews": {
                        "$ref": "#/components/schemas/properties/properties/Reviews"
                    },
                    "reviewCount": {
                        "$ref": "#/components/schemas/properties/properties/Number"
                    },
                    "reviewAvg": {
                        "$ref": "#/components/schemas/properties/properties/Number"
                    }
                }
            },
...
        },
        "parameters": {
...
            "Categories": {
                "name": "categories",
                "in": "body",
                "description": "Specify the product category",
                "required": true,
                "schema": {
                    "$ref": "#/components/schemas/properties/properties/String"
                }
            }
        },
        "responses": {
...
            "404": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Reporting a search problem"
                                }
                            }
                        }
                    }
                }
            },
...
        }
    }
}